requirements.txt
slack-bolt>=1.18.0
APScheduler>=3.10.0
requests>=2.31.0
python-dotenv>=1.0.0
pytz>=2023.3
